# move implementation of Timedelta to tslibs.timedeltas