--- a/pandas/_libs/tslib.pxd
+++ b/pandas/_libs/tslib.pxd
@@ -1,8 +1,8 @@
 from numpy cimport ndarray, int64_t
 
 from tslibs.conversion cimport convert_to_tsobject
+from tslibs.timedeltas cimport convert_to_timedelta64
 
-cpdef convert_to_timedelta64(object, object)
 cdef bint _check_all_nulls(obj)
 
 cdef _to_i8(object val)