--- a/pandas/_libs/tslibs/timedeltas.pyx
+++ b/pandas/_libs/tslibs/timedeltas.pyx
@@ -546,11 +546,11 @@ cdef _timedelta_value_kwargs(dict kwargs):
 # serve as a C extension type that shadows the Python class, where we do any
 # heavy lifting.
 cdef class _Timedelta(timedelta):
-    # cdef readonly:
-    #     int64_t value     # nanoseconds
-    #     object freq       # frequency reference
-    #     bint is_populated # are my components populated
-    #     int64_t _sign, _d, _h, _m, _s, _ms, _us, _ns
+    cdef readonly:
+        int64_t value     # nanoseconds
+        object freq       # frequency reference
+        bint is_populated # are my components populated
+        int64_t _sign, _d, _h, _m, _s, _ms, _us, _ns
 
     # higher than np.ndarray and np.matrix
     __array_priority__ = 100