… pandas/core/indexes/datetimelike.py
… pandas/core/indexes/period.py
… pandas/core/indexes/timedeltas.py
… pandas/core/resample.py
… pandas/core/tools/timedeltas.py
… pandas/tseries/offsets.py
… setup.py