@gfyoung 2017-11-03
@jbrockmendel : Minor comments.  Also, let's `asv` to make sure as usual.