@jbrockmendel 2017-11-03
Several asv runs:

```
taskset 5 asv continuous -E virtualenv -f 1.1 master HEAD -b timeseries
[...]
       before           after         ratio
     [b4375bde]       [98e339a5]
+     19.7±0.04μs       22.0±0.1μs     1.12  timeseries.Offsets.time_timeseries_day_incr
-     16.5±0.06ms      14.9±0.08ms     0.90  timeseries.Iteration.time_iter_periodindex_preexit
-     7.62±0.01μs      6.81±0.01μs     0.89  timeseries.DatetimeIndex.time_timestamp_tzinfo_cons

       before           after         ratio
     [b4375bde]       [98e339a5]
+     2.53±0.01ms      2.87±0.03ms     1.13  timeseries.AsOf.time_asof_nan_single
+      19.9±0.1μs       22.1±0.1μs     1.11  timeseries.Offsets.time_timeseries_day_incr
-      32.9±0.1μs       28.0±0.1μs     0.85  timeseries.Offsets.time_custom_bday_cal_decr
-        137±10ms         35.8±3ms     0.26  timeseries.DatetimeIndex.time_dti_tz_factorize
-        174±30ms         36.8±3ms     0.21  timeseries.DatetimeIndex.time_dti_factorize

       before           after         ratio
     [b4375bde]       [98e339a5]
+         429±6ms            2.56s     5.96  timeseries.AsOfDataFrame.time_asof_nan
+        408±10ms            2.30s     5.64  timeseries.AsOfDataFrame.time_asof
-     56.3±0.05μs      50.6±0.08μs     0.90  timeseries.SemiMonthOffset.time_begin_decr
-     7.26±0.08μs      6.51±0.02μs     0.90  timeseries.DatetimeIndex.time_timestamp_tzinfo_cons
-     21.9±0.09μs       18.7±0.1μs     0.86  timeseries.Offsets.time_custom_bday_apply_dt64
-        27.0±2ms       6.64±0.2ms     0.25  timeseries.DatetimeIndex.time_dti_tz_factorize
-        30.1±4ms       7.17±0.1ms     0.24  timeseries.DatetimeIndex.time_dti_factorize

       before           after         ratio
     [b4375bde]       [98e339a5]
+      17.1±0.2μs      20.1±0.08μs     1.18  timeseries.Offsets.time_custom_bday_apply_dt64
+      27.5±0.1μs       32.1±0.2μs     1.17  timeseries.Offsets.time_custom_bday_decr
+      18.2±0.2μs      20.1±0.06μs     1.11  timeseries.Offsets.time_timeseries_day_apply
-      20.6±0.1μs      18.7±0.06μs     0.91  timeseries.Offsets.time_custom_bday_incr
-     7.39±0.02μs      6.71±0.02μs     0.91  timeseries.DatetimeIndex.time_timestamp_tzinfo_cons
-     31.1±0.07μs       27.4±0.1μs     0.88  timeseries.Offsets.time_custom_bday_cal_incr_neg_n
-        2.77±0ms         1.68±0ms     0.61  timeseries.DatetimeIndex.time_add_timedelta
-         220±6ms       27.5±0.7ms     0.13  timeseries.AsOfDataFrame.time_asof_nan
-         205±3ms      24.2±0.05ms     0.12  timeseries.AsOfDataFrame.time_asof

      before           after         ratio
     [b4375bde]       [98e339a5]
+     2.72±0.01ms      3.31±0.09ms     1.22  timeseries.ToDatetime.time_iso8601_nosep
+     22.3±0.07μs       24.7±0.1μs     1.11  timeseries.AsOf.time_asof_single
+     2.50±0.03ms      2.75±0.02ms     1.10  timeseries.AsOf.time_asof_nan_single
-         443±3ms          397±2ms     0.90  timeseries.SemiMonthOffset.time_begin_incr_rng
-      8.65±0.1μs      7.72±0.04μs     0.89  timeseries.Offsets.time_timeseries_year_apply
-      7.80±0.3μs      6.95±0.03μs     0.89  timeseries.DatetimeIndex.time_timestamp_tzinfo_cons
```