@jreback 2017-11-04
you have some wild variability. pls run for timedeltas too.