@jbrockmendel 2017-11-04
> you have some wild variability

This is a truthfact.

> pls run for timedeltas too

```
asv continuous -f 1.1 -E virtualenv master HEAD -b timedelta
[...]
BENCHMARKS NOT SIGNIFICANTLY CHANGED

asv continuous -f 1.1 -E virtualenv master HEAD -b timedelta
[...]
       before           after         ratio
     [b4375bde]       [98e339a5]
+        59.2±7ms          120±6ms     2.03  inference.DtypeInfer.time_timedelta64_2

asv continuous -f 1.1 -E virtualenv master HEAD -b timedelta
[...]
BENCHMARKS NOT SIGNIFICANTLY CHANGED.

asv continuous -f 1.1 -E virtualenv master HEAD -b timedelta
[...]
       before           after         ratio
     [b4375bde]       [98e339a5]
-         106±2ms         84.4±2ms     0.80  inference.DtypeInfer.time_timedelta64_1

```