@jreback 2017-11-04
ok so it doesn't change perf, which is expected, good.