@jreback 2017-11-05
I think you need to add timedeltas.pxd to period in setup.py; I don't think you can remove anything from tslib def (but check). rebase.