@jbrockmendel 2017-11-07
If we can push this through we can port most of Timestamp to its own module.  Along with #18086 we can then get the rest.