@jbrockmendel 2017-11-08
> fix this too

Get rid of the matching one in `tslib` while we're at it?