@jreback 2017-11-08
thanks @jbrockmendel making nice progress!